api/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
flask-restx==1.0.5
google-api-core==2.11.0
google-auth==2.16.0
google-auth-oauthlib==0.8.0
google-cloud-core==2.3.2
google-cloud-firestore==2.9.0
googleapis-common-protos==1.58.0
gunicorn==20.1.0
oauthlib==3.2.2